requirements.txt
fastapi[standard]
uvicorn[standard]
langchain
langchain-google-genai
python-dotenv
